requirements.txt
asgiref==3.4.1
attrs==21.2.0
dj-database-url==0.5.0
Django==3.2.5
django-debug-toolbar==3.2.1
gunicorn==20.1.0
importlib-metadata==4.6.1
iniconfig==1.1.1
packaging==21.0
pluggy==0.13.1
psycopg2==2.9.1
psycopg2-binary==2.9.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.4
pytz==2021.1
reverse==0.1.0
sqlparse==0.4.1
toml==0.10.2
typing-extensions==3.10.0.0
whitenoise==5.3.0
zipp==3.5.0

djangorestframework~=3.12.4
selenium~=3.141.0